Move arrow_status.h.
